requirements.txt
pynput
cryptography
Pillow
python-dotenv